requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
pandas
tantivy
yfinance
groq 